third_party/WebKit/LayoutTests/fast/events/mouse-click-events-pseudo-element-expected.txt

CONSOLE ERROR: line 36: Uncaught SyntaxError: Illegal return statement
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
